Show missing values in 2d histograms; improved user manual; show help links
