+++ .gitignore
@@ -33,32 +33,28 @@ apache-tomcat-*
 web/classes
 
 # Generated by webpack
-web/src/main/webapp/d3-modules.js
+web/src/main/webapp/ui/*.js
+*.js.map
+
 web/src/main/webapp/columnConverter.js
 web/src/main/webapp/controlMenu.js
 web/src/main/webapp/heavyhittersview.js
 web/src/main/webapp/categoryCache.js
 web/src/main/webapp/tableData.js
 web/src/main/webapp/bundle.js
-web/src/main/webapp/errReporter.js
 web/src/main/webapp/heatMap.js
 web/src/main/webapp/heatMapArray.js
-web/src/main/webapp/vis.js
 web/src/main/webapp/hillview.js
 web/src/main/webapp/histogram.js
 web/src/main/webapp/initialObject.js
 web/src/main/webapp/histogram2d.js
 web/src/main/webapp/histogramBase.js
-web/src/main/webapp/menu.js
 web/src/main/webapp/rpc.js
 web/src/main/webapp/table.js
-web/src/main/webapp/ui.js
 web/src/main/webapp/util.js
-web/src/main/webapp/dialog.js
 web/src/main/webapp/pca.js
 web/src/main/webapp/lamp.js
 web/src/main/webapp/equalityFilter.js
-web/src/main/webapp/*.js.map
 web/src/main/webapp/node_modules/
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
