+++ web/.idea/dictionaries/mbudiu.xml
@@ -1,6 +1,7 @@
 <component name="ProjectDictionaryState">
   <dictionary name="mbudiu">
     <words>
+      <w>resized</w>
       <w>spdx</w>
     </words>
   </dictionary>
