--- web/src/main/webapp/d3-modules.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2017 VMware Inc. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * This file is just used to package together several d3 modules.
- */
-
-import d3 = require('d3');
-
-/*
-import d3drag = require('d3-drag');
-import d3select = require('d3-selection');
-import d3format = require('d3-format');
-import d3scale = require('d3-scale');
-import d3axis = require('d3-axis');
-import d3shape = require('d3-shape');
-import d3time = require('d3-time');
-export { d3drag, d3select, d3format, d3scale, d3axis, d3shape, d3time };
-*/
-
-export { d3 };
-
